Add MonadShell DSL
