+++ src/ShellRun/Types/Legend.hs
@@ -1,0 +1,15 @@
+module ShellRun.Types.Legend
+  ( LegendErr (..),
+    LegendMap,
+  )
+where
+
+import Data.Map.Strict (Map)
+import Data.Text (Text)
+
+data LegendErr
+  = FileErr Text
+  | ParseErr Text
+  deriving (Show)
+
+type LegendMap = Map Text Text
