--- src/ShellRun/Types/LegendMap.hs
@@ -1,9 +1,0 @@
-module ShellRun.Types.LegendMap
-  ( LegendMap,
-  )
-where
-
-import Data.Map.Strict (Map)
-import Data.Text (Text)
-
-type LegendMap = Map Text Text
